feat(layout): merge header and tab bar into a single row
The tab bar previously sat on its own row below the header, leaving a
large band of empty vertical space above it. Integrate the tabs into the
header so brand, tabs, and controls share one row, reclaiming that space.

- AppLayout now owns the Tabs.Root, rendering the tab list inside the
  header and the tab content in main; TabNavigation is removed and the
  TabItem type moves to AppLayout.
- Tabs live in a flex-1, horizontally scrollable region so they never
  push out the connection/language/theme controls.
- Responsive collapse: mobile shows icon-only tabs + compact controls,
  tablet adds tab labels and the connection status, desktop adds the
  brand text and the Disconnect label.
- Implement the previously-undefined scrollbar-none utility so the tab
  row scrolls without a visible scrollbar.
